+++ integrations/metadata-cache/index.ts
@@ -1,0 +1,1 @@
+export { MetadataIntegration } from './metadata-integration';
