+++ main.ts
@@ -11,9 +11,8 @@ import { MetadataIntegration } from './integrations/metadataCache';
 import { WorkspaceIntegration } from './integrations/workspace';
 
 // Importar componentes UI
-import { StatusBarController } from 'components/statusBar/status-bar-controller';
-import { StatusPaneView } from 'ui/components/status-pane-view';
 import { NoteStatusSettingTab } from 'settings/settings-tab';
+import { StatusBarController } from 'components/status-bar/status-bar-controller';
 
 export default class NoteStatus extends Plugin {
   settings: NoteStatusSettings;
@@ -38,24 +37,24 @@ export default class NoteStatus extends Plugin {
       await this.loadSettings();
       
       // 1. Inicializar servicios básicos
-      this.initializeServices();
+      // this.initializeServices();
       
       // 2. Registrar vistas y componentes UI
-      this.registerViews();
+      // this.registerViews();
       
       // 3. Inicializar integraciones
-      this.initializeIntegrations();
+      // this.initializeIntegrations();
       
       // 4. Registrar comandos
-      this.registerCommands();
+      // this.registerCommands();
       
       // 5. Registrar eventos personalizados
-      this.setupCustomEvents();
+      // this.setupCustomEvents();
       
       // 6. Esperar a que el layout esté listo para inicializar UI
-      this.app.workspace.onLayoutReady(() => {
-        this.initializeUI();
-      });
+      // this.app.workspace.onLayoutReady(() => {
+      //   this.initializeUI();
+      // });
       
     } catch (error) {
       console.error('Error loading Note Status plugin:', error);
@@ -68,227 +67,224 @@ export default class NoteStatus extends Plugin {
   }
 
   private initializeServices() {
-    this.statusService = new StatusService(this.app, this.settings);
-    this.styleService = new StyleService(this.settings);
+    // this.statusService = new StatusService(this.app, this.settings);
+    // this.styleService = new StyleService(this.settings);
   }
 
   private registerViews() {
     // Registrar vista del panel de estado
-    this.registerView('status-pane', (leaf) => {
-      return new StatusPaneView(leaf, this);
-    });
+    // this.registerView('status-pane', (leaf) => {
+    //   return new StatusPaneView(leaf, this);
+    // });
 
     // Añadir ícono en la barra lateral
-    this.addRibbonIcon('status-pane', 'Open status pane', () => {
-      this.openStatusPane();
-    });
+    // this.addRibbonIcon('status-pane', 'Open status pane', () => {
+    //   this.openStatusPane();
+    // });
     
     // Añadir pestaña de configuración
-    this.addSettingTab(new NoteStatusSettingTab(this.app, this, this.statusService));
+    // this.addSettingTab(new NoteStatusSettingTab(this.app, this, this.statusService));
   }
 
   private initializeIntegrations() {
     // Crear integraciones en orden de dependencia
     
-    // 1. Integraciones básicas primero
-    this.explorerIntegration = new ExplorerIntegration(this.app, this.settings, this.statusService);
-    this.toolbarIntegration = new ToolbarIntegration(this.app, this.settings, this.statusService);
+    // // 1. Integraciones básicas primero
+    // this.explorerIntegration = new ExplorerIntegration(this.app, this.settings, this.statusService);
+    // this.toolbarIntegration = new ToolbarIntegration(this.app, this.settings, this.statusService);
     
-    // 2. Integraciones que dependen de otras
-    this.fileMenuIntegration = new FileMenuIntegration(
-      this.app, 
-      this.settings, 
-      this.statusService,
-      this.explorerIntegration
-    );
+    // // 2. Integraciones que dependen de otras
+    // this.fileMenuIntegration = new FileMenuIntegration(
+    //   this.app, 
+    //   this.settings, 
+    //   this.statusService,
+    //   this.explorerIntegration
+    // );
     
-    this.editorIntegration = new EditorIntegration(this.app, this.settings, this.statusService);
+    // this.editorIntegration = new EditorIntegration(this.app, this.settings, this.statusService);
     
-    this.metadataIntegration = new MetadataIntegration(
-      this.app, 
-      this.settings, 
-      this.statusService,
-      this.explorerIntegration
-    );
+    // this.metadataIntegration = new MetadataIntegration(
+    //   this.app, 
+    //   this.settings, 
+    //   this.statusService,
+    //   this.explorerIntegration
+    // );
     
-    this.workspaceIntegration = new WorkspaceIntegration(
-      this.app, 
-      this.settings,
-      this.statusService,
-      this.toolbarIntegration
-    );
+    // this.workspaceIntegration = new WorkspaceIntegration(
+    //   this.app, 
+    //   this.settings,
+    //   this.statusService,
+    //   this.toolbarIntegration
+    // );
     
-    // 3. Registrar eventos en cada integración
-    this.fileMenuIntegration.registerFileMenus();
-    this.editorIntegration.registerEditorMenus();
-    this.metadataIntegration.registerMetadataEvents();
-    this.workspaceIntegration.registerWorkspaceEvents();
+    // // 3. Registrar eventos en cada integración
+    // this.fileMenuIntegration.registerFileMenus();
+    // this.editorIntegration.registerEditorMenus();
+    // this.metadataIntegration.registerMetadataEvents();
+    // this.workspaceIntegration.registerWorkspaceEvents();
   }
 
   private registerCommands() {
-    // Comando para actualizar estado
-    this.addCommand({
-      id: 'refresh-status',
-      name: 'Refresh status',
-      callback: () => {
-        this.refreshStatus();
-        new Notice('Note status refreshed!');
-      }
-    });
+    // // Comando para actualizar estado
+    // this.addCommand({
+    //   id: 'refresh-status',
+    //   name: 'Refresh status',
+    //   callback: () => {
+    //     this.refreshStatus();
+    //     new Notice('Note status refreshed!');
+    //   }
+    // });
 
-    // Comando para forzar actualización de UI
-    this.addCommand({
-      id: 'force-refresh-ui',
-      name: 'Force refresh user interface',
-      callback: () => this.forceRefreshUI()
-    });
+    // // Comando para forzar actualización de UI
+    // this.addCommand({
+    //   id: 'force-refresh-ui',
+    //   name: 'Force refresh user interface',
+    //   callback: () => this.forceRefreshUI()
+    // });
 
-    // Comando para insertar metadatos de estado
-    this.addCommand({
-      id: 'insert-status-metadata',
-      name: 'Insert status metadata',
-      editorCallback: (editor) => {
-        this.editorIntegration.insertStatusMetadata(editor);
-        new Notice('Status metadata inserted');
-      }
-    });
+    // // Comando para insertar metadatos de estado
+    // this.addCommand({
+    //   id: 'insert-status-metadata',
+    //   name: 'Insert status metadata',
+    //   editorCallback: (editor) => {
+    //     this.editorIntegration.insertStatusMetadata(editor);
+    //     new Notice('Status metadata inserted');
+    //   }
+    // });
 
-    // Comando para abrir panel de estado
-    this.addCommand({
-      id: 'open-status-pane',
-      name: 'Open status pane',
-      callback: () => this.openStatusPane()
-    });
+    // // Comando para abrir panel de estado
+    // this.addCommand({
+    //   id: 'open-status-pane',
+    //   name: 'Open status pane',
+    //   callback: () => this.openStatusPane()
+    // });
   }
 
   private setupCustomEvents() {
-    // Evento para cambios de configuración
-    window.addEventListener('note-status:settings-changed', this.saveSettings.bind(this));
+    // // Evento para cambios de configuración
+    // window.addEventListener('note-status:settings-changed', this.saveSettings.bind(this));
     
-    // Evento para cambios de estado
-    window.addEventListener('note-status:status-changed', this.handleStatusChanged.bind(this));
+    // // Evento para cambios de estado
+    // window.addEventListener('note-status:status-changed', this.handleStatusChanged.bind(this));
     
-    // Evento para actualización de UI
-    window.addEventListener('note-status:refresh-ui', this.refreshUI.bind(this));
+    // // Evento para actualización de UI
+    // window.addEventListener('note-status:refresh-ui', this.refreshUI.bind(this));
   }
 
   private initializeUI() {
-    // Inicializar barra de estado
-    this.statusBarController = new StatusBarController(
-      this.addStatusBarItem(), 
-      this.settings, 
-      this.statusService
-    );
+    // // Inicializar barra de estado
+    // this.statusBarController = new StatusBarController(
+    //   this.addStatusBarItem(), 
+    //   this.settings, 
+    //   this.statusService
+    // );
     
-    // Inicializar iconos del explorador (con retraso para evitar ralentizar el inicio)
-    if (this.settings.showStatusIconsInExplorer) {
-      setTimeout(() => {
-        this.explorerIntegration.updateAllFileExplorerIcons();
-      }, 2000);
-    }
+    // // Inicializar iconos del explorador (con retraso para evitar ralentizar el inicio)
+    // if (this.settings.showStatusIconsInExplorer) {
+    //   setTimeout(() => {
+    //     this.explorerIntegration.updateAllFileExplorerIcons();
+    //   }, 2000);
+    // }
   }
 
   private handleStatusChanged(event: CustomEvent) {
-    const { statuses, file } = event.detail;
+    // const { statuses, file } = event.detail;
     
-    // Actualizar barra de estado
-    this.statusBarController.update(statuses);
+    // // Actualizar barra de estado
+    // this.statusBarController.update(statuses);
     
-    // Actualizar toolbar
-    this.toolbarIntegration.updateStatusDisplay(statuses);
+    // // Actualizar toolbar
+    // this.toolbarIntegration.updateStatusDisplay(statuses);
     
-    // Actualizar explorador si es necesario
-    if (file) {
-      const fileObj = this.app.vault.getFileByPath(file);
-      if (fileObj) {
-        this.explorerIntegration.updateFileExplorerIcons(fileObj);
-      }
-    }
+    // // Actualizar explorador si es necesario
+    // if (file) {
+    //   const fileObj = this.app.vault.getFileByPath(file);
+    //   if (fileObj) {
+    //     this.explorerIntegration.updateFileExplorerIcons(fileObj);
+    //   }
+    // }
   }
 
   private refreshStatus() {
-    const activeFile = this.app.workspace.getActiveFile();
-    if (activeFile) {
-      const statuses = this.statusService.getFileStatuses(activeFile);
-      window.dispatchEvent(new CustomEvent('note-status:status-changed', { 
-        detail: { statuses, file: activeFile.path } 
-      }));
-    }
+    // const activeFile = this.app.workspace.getActiveFile();
+    // if (activeFile) {
+    //   const statuses = this.statusService.getFileStatuses(activeFile);
+    //   window.dispatchEvent(new CustomEvent('note-status:status-changed', { 
+    //     detail: { statuses, file: activeFile.path } 
+    //   }));
+    // }
   }
 
   private refreshUI() {
-    this.refreshStatus();
-    
-    // Actualizar otros componentes según sea necesario
-    // ...
+    // this.refreshStatus();
   }
 
   private forceRefreshUI() {
-    // Actualizar todos los elementos visibles
-    this.refreshStatus();
+    // // Actualizar todos los elementos visibles
+    // this.refreshStatus();
     
-    if (this.settings.showStatusIconsInExplorer) {
-      this.explorerIntegration.updateAllFileExplorerIcons();
-    }
+    // if (this.settings.showStatusIconsInExplorer) {
+    //   this.explorerIntegration.updateAllFileExplorerIcons();
+    // }
     
-    window.dispatchEvent(new CustomEvent('note-status:update-pane'));
+    // window.dispatchEvent(new CustomEvent('note-status:update-pane'));
     
-    new Notice('UI forcefully refreshed');
+    // new Notice('UI forcefully refreshed');
   }
 
   private async openStatusPane() {
-    // Comprobar si ya está abierto
-    const existing = this.app.workspace.getLeavesOfType('status-pane')[0];
-    if (existing) {
-      this.app.workspace.setActiveLeaf(existing);
-      return;
-    }
+    // // Comprobar si ya está abierto
+    // const existing = this.app.workspace.getLeavesOfType('status-pane')[0];
+    // if (existing) {
+    //   this.app.workspace.setActiveLeaf(existing);
+    //   return;
+    // }
     
-    // Crear una nueva hoja
-    const leaf = this.app.workspace.getLeftLeaf(false);
-    if (leaf) {
-      await leaf.setViewState({ type: 'status-pane', active: true });
-    }
+    // // Crear una nueva hoja
+    // const leaf = this.app.workspace.getLeftLeaf(false);
+    // if (leaf) {
+    //   await leaf.setViewState({ type: 'status-pane', active: true });
+    // }
   }
 
   async saveSettings() {
-    await this.saveData(this.settings);
+    // await this.saveData(this.settings);
     
-    // Actualizar servicios
-    this.statusService.updateSettings(this.settings);
-    this.styleService.updateSettings(this.settings);
+    // // Actualizar servicios
+    // this.statusService.updateSettings(this.settings);
+    // this.styleService.updateSettings(this.settings);
     
-    // Actualizar integraciones
-    this.updateIntegrationsSettings();
+    // // Actualizar integraciones
+    // this.updateIntegrationsSettings();
   }
 
   private updateIntegrationsSettings() {
-    // Actualizar configuración en todas las integraciones
-    this.explorerIntegration.updateSettings(this.settings);
-    this.fileMenuIntegration.updateSettings(this.settings);
-    this.editorIntegration.updateSettings(this.settings);
-    this.toolbarIntegration.updateSettings(this.settings);
-    this.metadataIntegration.updateSettings(this.settings);
-    this.workspaceIntegration.updateSettings(this.settings);
+    // // Actualizar configuración en todas las integraciones
+    // this.explorerIntegration.updateSettings(this.settings);
+    // this.fileMenuIntegration.updateSettings(this.settings);
+    // this.editorIntegration.updateSettings(this.settings);
+    // this.toolbarIntegration.updateSettings(this.settings);
+    // this.metadataIntegration.updateSettings(this.settings);
+    // this.workspaceIntegration.updateSettings(this.settings);
     
-    // Actualizar componentes UI
-    this.statusBarController.updateSettings(this.settings);
+    // // Actualizar componentes UI
+    // this.statusBarController.updateSettings(this.settings);
   }
 
   onunload() {
-    // Limpiar eventos personalizados
-    window.removeEventListener('note-status:settings-changed', this.saveSettings.bind(this));
-    window.removeEventListener('note-status:status-changed', this.handleStatusChanged.bind(this));
-    window.removeEventListener('note-status:refresh-ui', this.refreshUI.bind(this));
+    // // Limpiar eventos personalizados
+    // window.removeEventListener('note-status:settings-changed', this.saveSettings.bind(this));
+    // window.removeEventListener('note-status:status-changed', this.handleStatusChanged.bind(this));
+    // window.removeEventListener('note-status:refresh-ui', this.refreshUI.bind(this));
     
-    // Limpiar integraciones
-    this.explorerIntegration.unload();
-    this.toolbarIntegration.unload();
+    // // Limpiar integraciones
+    // this.explorerIntegration.unload();
+    // this.toolbarIntegration.unload();
     
-    // Limpiar servicios
-    this.styleService.unload();
+    // // Limpiar servicios
+    // this.styleService.unload();
     
-    // Limpiar componentes UI
-    this.statusBarController.unload();
+    // // Limpiar componentes UI
+    // this.statusBarController.unload();
   }
 }
